% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildSPSignature.R
\name{addSpacesToIniFile}
\alias{addSpacesToIniFile}
\title{addSpacesToIniFile}
\usage{
addSpacesToIniFile(filename)
}
\arguments{
\item{filename}{The file to read from and write to.}
}
\description{
replace "=" by " = " in a file.
}
