% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildCoxElasticNet.R
\name{buildCoxElasticNet}
\alias{buildCoxElasticNet}
\title{buildCoxElasticNet}
\usage{
buildCoxElasticNet(msa, iterations)
}
\arguments{
\item{msa}{The msaWrapper object to work with.}
}
\description{
Generic function.
}
