% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildCoxElasticNet.R, R/buildRandomForest.R
\name{buildRandomForest.msaWrapperOclass}
\alias{buildRandomForest.msaWrapperOclass}
\title{buildCoxElasticNet.msaWrapperOclass}
\usage{
\method{buildRandomForest}{msaWrapperOclass}(msa, iterations = 200, byOOB = TRUE)

\method{buildRandomForest}{msaWrapperOclass}(msa, iterations = 200, byOOB = TRUE)
}
\arguments{
\item{msa}{The msaWrapper object to work with.}

\item{iterations}{The number of randomForest calls to make for the active covars.}

\item{byOOB}{Use out of bag performance measure, if false use 50:50 cross validation}
}
\value{
An object containing the trained model and it's performance.

An object containing the trained RF and it's performance.
}
\description{
Build a Cox Elastic Net regularized model for ordinal class outcome

Build a Random Forest Classifier for ordinal class outcome
}
