% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildRandomForest.R
\name{buildRandomForest.msaWrapperTte}
\alias{buildRandomForest.msaWrapperTte}
\title{buildRandomForest.msaWrapperTte}
\usage{
\method{buildRandomForest}{msaWrapperTte}(msa, iterations = 200, byOOB = TRUE)
}
\arguments{
\item{msa}{The msaWrapper object to work with.}

\item{iterations}{The number of randomForest calls to make for the active covars.}

\item{byOOB}{Use out of bag performance measure, if false use 50:50 cross validation}
}
\value{
An object containing the trained RF and it's performance.
}
\description{
Build a Random Forest Classifier for time to event outcome
}
