% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildRandomForest.R
\name{buildRandomForest}
\alias{buildRandomForest}
\title{buildRandomForest}
\usage{
buildRandomForest(msa, iterations, byOOB)
}
\arguments{
\item{msa}{The msaWrapper object to work with.}
}
\description{
Generic function.
}
