% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildSPSignature.R
\name{buildSPSignature.msaWrapperOclass}
\alias{buildSPSignature.msaWrapperOclass}
\title{buildSPSignature.msaWrapperOclass}
\usage{
\method{buildSPSignature}{msaWrapperOclass}(msa, iterations = 200)
}
\arguments{
\item{msa}{The msaWrapper object to work with.}

\item{iterations}{The number of cross validation iterations to make.}
}
\value{
An object containing the signature and it's performance.
}
\description{
Build a Saddle Point Signature for ordinal class outcome.
This requires the installation and valid license for that software.
}
