% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildSPSignature.R
\name{buildSPSignature.msaWrapperTte}
\alias{buildSPSignature.msaWrapperTte}
\title{buildSPSignature.msaWrapperTte}
\usage{
\method{buildSPSignature}{msaWrapperTte}(msa, runName, iterations = 200)
}
\arguments{
\item{msa}{The msaWrapper object to work with.}

\item{iterations}{The number of cross validation iterations to make.}
}
\value{
An object containing the signature.
}
\description{
Build a Saddle Point Signature for time to event outcome.
This requires the installation and valid license for that software.
}
