% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildSPSignature.R
\name{buildSPSignature}
\alias{buildSPSignature}
\title{buildSPSignature}
\usage{
buildSPSignature(msa, runName, iterations)
}
\arguments{
\item{msa}{The msaWrapper object to work with.}
}
\description{
Generic function.
}
