% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildSPSignature.R
\name{createTemplateIni_BatchRegression}
\alias{createTemplateIni_BatchRegression}
\title{createTemplateIni_BatchRegression}
\usage{
createTemplateIni_BatchRegression(filename)
}
\arguments{
\item{filename}{The filename to write to.}
}
\description{
Create a template SPS ini file for batch regression.
}
