% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildSPSignature.R
\name{export_SPS_file}
\alias{export_SPS_file}
\title{export_SPS_file}
\usage{
export_SPS_file(export_dat, export_name, type, C = NA, R = NA)
}
\arguments{
\item{export_dat}{Data frame of covariates and outcome (last columns), which may have sample identifiers as the first column.}

\item{export_name}{A name for the data to be used for file names.}

\item{type}{A Saddle Point data type. See description.}

\item{C}{For ordinal class data, the number of classes.}

\item{R}{For time to event data, the number of event types.}
}
\description{
Export a data frame to .data and .names files for use with Saddle Point software
}
\details{
Saddle Point data types
1: time-to-event outcome data with sample identifiers
2: time-to-event outcome data without sample identifiers
3: ordinal class outcome data with sample identifiers
4: ordinal class outcome data without sample identifiers
5: general real outcome data with sample identifiers
6: general real outcome data without sample identifiers
}
