% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomeCorrelations.R
\name{generateOutcomeCorrelations}
\alias{generateOutcomeCorrelations}
\title{generateOutcomeCorrelations}
\usage{
generateOutcomeCorrelations(msa, method = c("kendall", "pearson", "spearman"))
}
\arguments{
\item{msa}{The msaWrapper object to work with.}

\item{method}{Correlation method to use.}

\item{textSize}{Data labels text size.}

\item{groups}{To restrict covars to those in certain groups.}
}
\description{
This function correlates all covars against outcome.
It places the results in the msaWrapper.
Usually use kendall rank correlation for tte and ordinal class.
For tte data, use only the events, ignoring censored data.
}
