% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildElasticNet.R
\name{getElasNetSignatures}
\alias{getElasNetSignatures}
\title{getElasNetSignatures}
\usage{
getElasNetSignatures(elasnetModel, is_optimal_signature = FALSE)
}
\arguments{
\item{elasnetModel}{The buildElasticNet.msaWrapperOclass object built from buildElasticNet}

\item{is_optimal_signature}{optimal_signature =True to get optimal features, else return strict signatures}
}
\value{
The Riskscore formula
}
\description{
getElasNetSignatures
}
