% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictElasticNet.R
\name{getElasticNetPerformance}
\alias{getElasticNetPerformance}
\title{getElasticNetPerformance}
\usage{
getElasticNetPerformance(predictResults, Outcome)
}
\arguments{
\item{predictResults}{Test data to predict with.}

\item{Outcome}{Test data to predict with.}
}
\value{
return AUC
}
\description{
getElasticNetPerformance
}
