% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildSPSignature.R
\name{getNormalisedSignatures}
\alias{getNormalisedSignatures}
\title{getNormalisedSignatures}
\usage{
getNormalisedSignatures(runName, regressionFolder = Reg_SETCV_MAP_L2)
}
\arguments{
\item{runName}{Test data to predict with.}

\item{regressionFolder}{The regressionfolder saved in the training step.}
}
\value{
The Riskscore formula
}
\description{
getNormalisedSignatures
}
