% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictSPSignature.R
\name{getSPSignaturePerformance}
\alias{getSPSignaturePerformance}
\title{getSPSignaturePerformance}
\usage{
getSPSignaturePerformance(predictResults, Outcome)
}
\arguments{
\item{predictResults}{Test data to predict with.}

\item{Outcome}{Test data to predict with.}
}
\value{
return AUC
}
\description{
getSPSignaturePerformance
}
