% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msaWrapper-package.R
\docType{package}
\name{msaWrapper-package}
\alias{msaWrapper}
\alias{msaWrapper-package}
\title{msaWrapper: Multivariate Survival Analysis}
\description{
This package unifies the exploration and use of multivariate survival analysis methods for time to event and ordinal class data.
}
\author{
\strong{Maintainer}: Paul Barber \email{paul.barber@kcl.ac.uk} (\href{https://orcid.org/0000-0002-8595-1141}{ORCID})

}
\keyword{internal}
