% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{msaWrapperCreate}
\alias{msaWrapperCreate}
\title{msaWrapperCreate}
\usage{
msaWrapperCreate(data, outcome, rowsLabelled = F, group = NA, groupLabels = NA)
}
\arguments{
\item{data}{A data frame of numeric data (one hot encoded etc).}

\item{outcome}{A vector of classes (for ordinal class data), or a data frame (for time to event data) with 2 cols: time, event}

\item{rowsLabelled}{Indicates if data contains an ID as first column T/F.}

\item{group}{A vector of group labels for each covariate.}

\item{groupLabels}{A vector of names for the groups 1-G.}
}
\description{
This function creates a msaWrapper object that contains the data
and will contain future results..
}
