% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{msaWrapperPrint}
\alias{msaWrapperPrint}
\title{print.msaWrapperTte}
\usage{
msaWrapperPrint(msa)
}
\arguments{
\item{msa}{The msaWrapper object}
}
\description{
This function print info about a msaWrapper object.
This is the generic for msaWrapper.
See overloaded print functions that are exported below.
}
