% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSPSignature.R
\name{plotBetaTable}
\alias{plotBetaTable}
\title{plotBetaTable}
\usage{
plotBetaTable(
  results.folder,
  cex = 0.5,
  rowspacing = 0.5,
  col = 1,
  pch = 18,
  xmin = -2.9,
  xmax = 3.6,
  betaTicks = c(-1, -0.5, 0, 0.5, 1),
  covariatepos = xmin,
  classpos = -0.8,
  betapos = 2.1,
  pvaluepos = 3.2,
  weightpos = 4,
  rankpos = 4.7,
  useClass = FALSE,
  useWeight = FALSE,
  useRank = FALSE
)
}
\arguments{
\item{results.folder}{Result folder to get required file}

\item{cex}{Text font}
}
\description{
Plot of Betas through the optimisation of SP Signature batch regression using betas_with_errors_optimised.txt
}
\details{
Required from the parent folder of data.folder
betas_with_errors_optimised.txt
}
