% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCoxElasticNet.R
\name{plotCoxElasticNetCoefs}
\alias{plotCoxElasticNetCoefs}
\title{plotCoxElasticNetCoefs
Plot the coefficients as a function of Lambda,}
\usage{
plotCoxElasticNetCoefs(cen)
}
\arguments{
\item{cen}{The msaWrapperCoxElasticNet object from buildCoxElasticNet().}
}
\description{
plotCoxElasticNetCoefs
Plot the coefficients as a function of Lambda,
}
