% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCoxElasticNet.R
\name{plotCoxElasticNetPerformance}
\alias{plotCoxElasticNetPerformance}
\title{plotCoxElasticNetPerformance
Plot the performance as a function of Lambda,}
\usage{
plotCoxElasticNetPerformance(cen)
}
\arguments{
\item{cen}{The msaWrapperCoxElasticNet object from buildCoxElasticNet().}
}
\description{
plotCoxElasticNetPerformance
Plot the performance as a function of Lambda,
}
