% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{plotDataCrossCorrelations}
\alias{plotDataCrossCorrelations}
\title{plotDataCrossCorrelations}
\usage{
plotDataCrossCorrelations(
  msa,
  method = c("kendall", "pearson", "spearman"),
  textSize = 1,
  group1,
  group2
)
}
\arguments{
\item{msa}{The msaWrapper object to work with.}

\item{method}{Correlation method to use.}

\item{textSize}{Axes labels text size.}

\item{group1}{Vector of groups that define the first set of covars.}

\item{group2}{Vector of groups that define the second set of covars.}
}
\description{
This function plots a correlation matrix of 2 sets of covars from given groups.
}
