% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotElasticNet.R
\name{plotElasticNetCoefs}
\alias{plotElasticNetCoefs}
\title{plotElasticNetCoefs
Plot the coefficients as a function of Lambda/Norm/Deviance,}
\usage{
plotElasticNetCoefs(cen, xvar = "lambda")
}
\arguments{
\item{cen}{The msaWrapperElasticNet object from buildElasticNet().}

\item{xvar}{lambda/norm/dev to decide X axis}
}
\description{
plotElasticNetCoefs
Plot the coefficients as a function of Lambda/Norm/Deviance,
}
