% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotElasticNet.R
\name{plotElasticNetPerformance}
\alias{plotElasticNetPerformance}
\title{plotElasticNetPerformance
Plot the performance as a function of Lambda,}
\usage{
plotElasticNetPerformance(cen)
}
\arguments{
\item{cen}{The msaWrapperElasticNet object from buildElasticNet().}
}
\description{
plotElasticNetPerformance
Plot the performance as a function of Lambda,
}
