% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes.R
\name{plotOutcome.msaWrapperOclass}
\alias{plotOutcome.msaWrapperOclass}
\title{plotOutcome.msaWrapperOclass}
\usage{
\method{plotOutcome}{msaWrapperOclass}(msa)
}
\arguments{
\item{msa}{The msaWrapper object to work with.}
}
\description{
Plot bar chart of outcome classes for an ordinal class outcome.
}
