% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes.R
\name{plotOutcome.msaWrapperTte}
\alias{plotOutcome.msaWrapperTte}
\title{plotOutcome.msaWrapperTte}
\usage{
\method{plotOutcome}{msaWrapperTte}(msa)
}
\arguments{
\item{msa}{The msaWrapper object to work with.}
}
\description{
Plot bar chart of outcome classes for time to event outcome.
}
