% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes.R
\name{plotOutcomeByClass.msaWrapperOclass}
\alias{plotOutcomeByClass.msaWrapperOclass}
\title{plotOutcomeByClass.msaWrapperOclass}
\usage{
\method{plotOutcomeByClass}{msaWrapperOclass}(msa, classVector, strataName)
}
\arguments{
\item{msa}{The msaWrapper object to work with.}

\item{classVector}{Vector of factors that identify the classes.}

\item{strataName}{A string name for the class division.}
}
\description{
Plot bar chart of outcome classes for an ordinal class outcome.
}
