% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes.R
\name{plotOutcomeByClass}
\alias{plotOutcomeByClass}
\title{plotOutcomeByClass}
\usage{
plotOutcomeByClass(msa, classVector, strataName)
}
\arguments{
\item{msa}{The msaWrapper object to work with.}
}
\description{
Generic function.
}
