% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomeCorrelations.R
\name{plotOutcomeCorrelations}
\alias{plotOutcomeCorrelations}
\title{plotOutcomeCorrelations}
\usage{
plotOutcomeCorrelations(
  msa,
  method = c("kendall", "pearson", "spearman"),
  textSize = 3,
  groups = vector()
)
}
\arguments{
\item{msa}{The msaWrapper object to work with.}

\item{method}{Correlation method to use.}

\item{textSize}{Data labels text size.}

\item{groups}{To restrict covars to those in certain groups.}
}
\description{
This function plots a volcano plot all covars against outcome.
Usually use kendall rank correlation for tte and ordinal class.
For tte data, use only the events, ignoring censored data.
}
