% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRandomForest.R
\name{plotRandomForestImportanceSwimmers}
\alias{plotRandomForestImportanceSwimmers}
\title{plotRandomForestImportanceSwimmers}
\usage{
plotRandomForestImportanceSwimmers(
  RSF,
  show_best_performance_line = TRUE,
  title = "",
  text_size = NA
)
}
\arguments{
\item{RSF}{Object returned by buildRandomForest.msaWrapperTte}

\item{show_best_performance_line}{TRUE/FALSE to show or not this dotted line at the optimal number of covariates.}

\item{title}{An optional title for the plot.}
}
\description{
Plot of importance scores through the optimisation of RF batch regression.#'
}
