% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRandomForest.R
\name{plotRandomForestPerformance.msaWrapperRandomSurvivalForest}
\alias{plotRandomForestPerformance.msaWrapperRandomSurvivalForest}
\title{plotRandomForestPerformance.msaWrapperRandomSurvivalForest}
\usage{
\method{plotRandomForestPerformance}{msaWrapperRandomSurvivalForest}(rf)
}
\arguments{
\item{rf}{The RF object from buildRandomForest().}
}
\description{
Plot the training performance of a random forest model.
}
