% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRandomForest.R
\name{plotRandomForestPerformance.msaWrapperRFClassifier}
\alias{plotRandomForestPerformance.msaWrapperRFClassifier}
\title{plotRandomForestPerformance.msaWrapperRFClassifier}
\usage{
\method{plotRandomForestPerformance}{msaWrapperRFClassifier}(rf)
}
\arguments{
\item{rf}{The RF object from buildRandomForest().}
}
\description{
Plot the training performance of a random forest model.
}
