% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRandomForest.R
\name{plotRandomForestPerformance}
\alias{plotRandomForestPerformance}
\title{plotRandomForestPerformance
Generic fn.
Plot the training performance of a random forest model}
\usage{
plotRandomForestPerformance(rf)
}
\arguments{
\item{rf}{The RF object from buildRandomForest().}
}
\description{
plotRandomForestPerformance
Generic fn.
Plot the training performance of a random forest model
}
