% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSPSignature.R
\name{plotSPSignatureBetaSwimmers}
\alias{plotSPSignatureBetaSwimmers}
\title{plotSPSignatureBetaSwimmers}
\usage{
plotSPSignatureBetaSwimmers(
  SPSig,
  show_best_performance_line = TRUE,
  selectedSig = FALSE,
  textreducescale = 1
)
}
\arguments{
\item{SPSig}{Object returned by buildSPSignature.msaWrapperTte}

\item{show_best_performance_line}{TRUE/FALSE to show or not this dotted line at the optimal number of covariates.}

\item{selectedSig}{TRUE/FALSE to plot full or selected Signatures in the Plot;}

\item{textreducescale}{To reduce the Text size, pass value >1; Useful for lower number of covariates}
}
\description{
Plot of Betas through the optimisation of SP Signature batch regression.
}
\details{
Required from data folder:
<data.name>.names file

Required from the parent folder of data.folder
SPS_Signature_logfile.txt

Required from results.folder
Betas_optimised.txt
Beta_values.txt
}
