% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSPSignature.R
\name{plotSPSignatureOClass}
\alias{plotSPSignatureOClass}
\title{plotSPSignatureOClass}
\usage{
plotSPSignatureOClass(SPSig)
}
\arguments{
\item{SPSig}{Object returned by buildSPSignature.msaWrapperOclass}
}
\description{
Plot performance of ordinal class signature/risk score.
Required from the results folder
BetaZ_<runName>.dat
}
