% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSPSignature.R
\name{plotSPSignatureTte}
\alias{plotSPSignatureTte}
\title{plotSPSignatureTte}
\usage{
plotSPSignatureTte(SPSig)
}
\arguments{
\item{SPSig}{Object returned by buildSPSignature.msaWrapperTte}
}
\description{
Plot performance of tte signature/risk score. Print other stats.
Required from the results folder
BetaZ_<runName>.dat
}
