% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{plotStrongCorrelations}
\alias{plotStrongCorrelations}
\title{plotStrongCorrelations}
\usage{
plotStrongCorrelations(
  msa,
  s,
  method = c("kendall", "pearson", "spearman"),
  trans = "identity"
)
}
\arguments{
\item{msa}{The msaWrapper object to work with.}

\item{s}{A data frame with at least 2 cols that have pairs of covar names to plot. like that produced by strongDataCorrelations()}

\item{trans}{Any of scale_continuous, incl. "log10", "log2" etc.}
}
\description{
Plot strong correlations in the covars in data frame s.
}
