% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictSPSignature.R
\name{predictSPSignature}
\alias{predictSPSignature}
\title{predictSPSignature}
\usage{
predictSPSignature(testData, runName, regressionFolder = Reg_SETCV_MAP_L2)
}
\arguments{
\item{testData}{Test data to predict with.}

\item{runName}{Test data to predict with.}

\item{regressionFolder}{The regressionfolder saved in the training step.}
}
\value{
Predicted Results
}
\description{
predictSPSignature
}
