% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.msaWrapperOclass}
\alias{print.msaWrapperOclass}
\title{print.msaWrapperOclass}
\usage{
\method{print}{msaWrapperOclass}(msa, ...)
}
\arguments{
\item{msa}{The msaWrapper object}
}
\description{
This function print info about a msaWrapperOClass object
}
