% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.msaWrapperTte}
\alias{print.msaWrapperTte}
\title{print.msaWrapperTte}
\usage{
\method{print}{msaWrapperTte}(msa, ...)
}
\arguments{
\item{msa}{The msaWrapper object}
}
\description{
This function print info about a msaWrapperTte object
}
