% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataReduction.R
\name{reduceByCorrelationWithinGroups}
\alias{reduceByCorrelationWithinGroups}
\title{reduceByCorrelationWithinGroups}
\usage{
reduceByCorrelationWithinGroups(
  msa,
  method = c("kendall", "pearson", "spearman"),
  threshold = 0.6
)
}
\arguments{
\item{msa}{The msaWrapper object to work with.}

\item{method}{Correlation method to use.}
}
\value{
A new msaWrapper object with reduced data.
}
\description{
Reduce the number of covars by removing those that are strongly correlated
within the groups, retaining the one best correlated with outcome.
}
