% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{strongDataCorrelations}
\alias{strongDataCorrelations}
\title{strongDataCorrelations}
\usage{
strongDataCorrelations(
  msa,
  threshold = 0.95,
  method = c("kendall", "pearson", "spearman"),
  groups = vector()
)
}
\arguments{
\item{msa}{The msaWrapper object to work with.}

\item{threshold}{Threshold for a correlation to be considered strong.}

\item{method}{Correlation method to use.}

\item{groups}{To restrict covars to those in certain groups.}
}
\value{
data frame of strong correlations.
}
\description{
Find strong correlations in the covars
}
