% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildCoxElasticNet.R
\name{buildCoxElasticNet.msaWrapperTte}
\alias{buildCoxElasticNet.msaWrapperTte}
\title{buildCoxElasticNet.msaWrapperTte}
\usage{
\method{buildCoxElasticNet}{msaWrapperTte}(msa, iterations = 200)
}
\arguments{
\item{msa}{The msaWrapper object to work with.}

\item{iterations}{The number of cv.glmnet calls to make.}
}
\value{
An object containing the trained model and it's performance.
}
\description{
Build a Cox Elastic Net regularized model for ordinal class outcome.
Uses glmnet::cv.glmnet() for optimisation.
Uses 3-fold cross validation (the minimum glmnet allows).
Data are normalised to zero mean and unit sd.
Missing data are imputed with the mean.
Returns optimal values from the peak of cross validation performance,
and a 'strict' values from the most regularized model with
performance within 1 standard deviation of the peak value.
}
