% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomeCorrelations.R
\name{plotOutcomeCorrelations}
\alias{plotOutcomeCorrelations}
\title{plotOutcomeCorrelations}
\usage{
plotOutcomeCorrelations(
  msa,
  labelAlpha = 0.1,
  hlineAlpha = 0.05,
  textSize = 3,
  groups = vector()
)
}
\arguments{
\item{msa}{The msaWrapper object to work with.}

\item{labelAlpha}{Significance Value level to display Signature names}

\item{hlineAlpha}{Horizontal line significance Value leve}

\item{textSize}{Data labels text size.}

\item{groups}{To restrict covars to those in certain groups.}
}
\description{
This function plots a volcano plot all covars against outcome.
Must use generateOutcomeCorrelations first.
The values and method use are determined by generateOutcomeCorrelations.
}
