% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomeCorrelations.R
\name{plotOutcomeCorrelations}
\alias{plotOutcomeCorrelations}
\title{plotOutcomeCorrelations}
\usage{
plotOutcomeCorrelations(msa, textSize = 3, groups = vector())
}
\arguments{
\item{msa}{The msaWrapper object to work with.}

\item{textSize}{Data labels text size.}

\item{groups}{To restrict covars to those in certain groups.}
}
\description{
This function plots a volcano plot all covars against outcome.
Must use generateOutcomeCorrelations first.
The values and method use are determined by generateOutcomeCorrelations.
}
